requirements.txt
# Core dependencies for Career Application Agent

# Note: PocketFlow is included as pocketflow.py (100 lines)
# We import from it rather than re-implementing
# This allows easy updates by replacing the file

# YAML parsing for career database
PyYAML>=6.0.1

# JSON schema validation
jsonschema>=4.20.0

# LLM access through OpenRouter (using OpenAI SDK format)
openai>=1.0.0

# Web utilities (for future company research features)
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pydantic>=2.5.0
python-dateutil>=2.8.2

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
mypy>=1.8.0
ruff>=0.1.0

# Optional: For better async performance
uvloop>=0.19.0 ; platform_system != "Windows"